splitting gun auth
