+++ .gitignore
@@ -143,4 +143,6 @@ dist
 # SvelteKit build / generate output
 .svelte-kit
 
+note
+
 # End of https://www.toptal.com/developers/gitignore/api/node
