+++ client/src/App.js
@@ -1,66 +1,19 @@
 import "./App.css";
-import GUN from "gun/gun";
-import SEA from "gun/sea";
-import React from "react";
-import { useState } from "react";
+import React, { useState } from "react";
+import {
+  createUser,
+  authUser,
+  authPair,
+  authKeyPair,
+  logout,
+  check,
+} from "./hooks";
 
 export default function App() {
-  // initialize gun
-  const gun = GUN({
-    peers: ["http://localhost:3030/gun"],
-  });
-  // create object references
-  const user = gun.user();
-
   // states
   const [username, setUsername] = useState("");
   const [password, setPassword] = useState("");
   const [key, setKey] = useState("");
-
-  // hooks
-  function createUser() {
-    user.create(username, password, (ack) => {
-      console.log(ack);
-    });
-  }
-
-  function authUser() {
-    user.auth(username, password, (ack) => {
-      console.log("logged in, public key:", ack.get);
-    });
-  }
-
-  async function authPair() {
-    const pair = await SEA.pair((ack) =>
-      console.log("pair cb |", JSON.stringify(ack))
-    );
-    user.auth(pair, (ack) => {
-      setKey(JSON.stringify(pair));
-      console.log("success", ack);
-    });
-  }
-
-  function authKeyPair() {
-    user.auth(key, (ack) => console.log("success", ack));
-  }
-
-  function logout() {
-    user.leave();
-    if (user._.sea) {
-      console.log("not logout", user._.sea);
-    } else {
-      console.log("logged out", user._.sea);
-    }
-  }
-
-  function check() {
-    gun.on("auth", (ack) => console.log(ack));
-    if (user.is) {
-      console.log("loggedin, key= ", JSON.stringify(user.is));
-    } else {
-      console.log("not logged in", user.is);
-    }
-  }
 
   return (
     <div className="App">
@@ -85,42 +38,42 @@ export default function App() {
       <button
         className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline"
         type="button"
-        onClick={authUser}
+        onClick={() => authUser(username, password)}
       >
         Login
       </button>
       <button
         className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline"
         type="button"
-        onClick={authKeyPair}
+        onClick={() => authKeyPair(key)}
       >
         Login Key
       </button>
       <button
         className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline"
         type="button"
-        onClick={authPair}
+        onClick={() => authPair(setKey)}
       >
         Login Pair
       </button>
       <button
         className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline"
         type="button"
-        onClick={createUser}
+        onClick={() => createUser(username, password)}
       >
         Register
       </button>
       <button
         className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline"
         type="button"
-        onClick={logout}
+        onClick={() => logout()}
       >
         Logout
       </button>
       <button
         className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline"
         type="button"
-        onClick={check}
+        onClick={() => check()}
       >
         Check
       </button>
